remove deprecated numpy types

+++ xrt/backends/raycing/oes.py
@@ -2281,7 +2281,7 @@ class GeneralFZPin0YZ(OE):
         if self.minHalfLambda is None:
             self.minHalfLambda = halfLambda.min()
         halfLambda -= self.minHalfLambda + self.phaseShift - phi*self.vorticity
-        zone = np.ones_like(x, dtype=np.int) * (self.N+2)
+        zone = np.ones_like(x, dtype=np.int32) * (self.N+2)
         zone[good] = np.floor(halfLambda).astype(np.int)
 #        N = 0
 #        while N < self.N:
